+++ src/data_structures/disjoint_set/quick_find.rs
@@ -6,15 +6,15 @@ pub struct QuickFind {
 }
 
 impl QuickFind {
-    pub fn new(n: uint) -> QuickFind {
+    pub fn new(n: usize) -> QuickFind {
         QuickFind{ id: Vec::from_fn(n, |i| i as int) }
     }
 
-    pub fn connected(&self, p: uint, q: uint) -> bool {
+    pub fn connected(&self, p: usize, q: usize) -> bool {
         self.id[p] == self.id[q]
     }
 
-    pub fn union(&mut self, p: uint, q: uint) {
+    pub fn union(&mut self, p: usize, q: usize) {
         let pid = self.id[p];
         let qid = self.id[q];
         for i in range(0, self.id.len()) {
